+++ common/protocol/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/protocol@1.3.0...@kyvejs/protocol@1.3.1) (2025-01-15)
+
+**Note:** Version bump only for package @kyvejs/protocol
+
 # [1.3.0](https://github.com/KYVENetwork/kyvejs/compare/@kyvejs/protocol@1.2.4...@kyvejs/protocol@1.3.0) (2024-10-30)
 
 ### Features